Start splitting onoff cluster integrations.
We want ifdefs to be less split around. This improves
redability/maintainability.
